requirements.txt
flask
firebase-admin
pillow
gunicorn
